Moved post display logic to a PostContainer component which is rendered on FeedPage (so a NewPost component can be rendered above) Started working on basic CSS for displaying posts as cards
