+++ frontend/src/components/PostContainer/PostContainer.css
@@ -1,0 +1,33 @@
+.post-container {
+  display: flex;
+  flex-direction: column;
+  /* grid-template-columns: repeat(auto-fill, minmax(300px, 1fr)); */
+  gap: 20px;
+  padding: 20px;
+  max-width: 1200px;
+  margin: 0 auto;
+}
+
+/* Card styling for each post */
+.post-container > article {
+  background-color: rgba(240, 240, 240, 0.85); /* Slightly transparent grey */
+  border-radius: 8px; /* Subtly rounded corners */
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  padding: 20px;
+  transition:
+    transform 0.2s ease-in-out,
+    box-shadow 0.2s ease-in-out;
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+  .post-container {
+    /* grid-template-columns: repeat(auto-fill, minmax(250px, 1fr)); */
+    gap: 15px;
+    padding: 15px;
+  }
+
+  .post-container > div {
+    padding: 15px;
+  }
+}
