+++ frontend/src/pages/Feed/FeedPage.jsx
@@ -1,42 +1,13 @@
-import { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-
-import { getPosts } from "../../services/posts";
-import Post from "../../components/Post";
 import Header from "../../components/Header";
+import PostContainer from "../../components/PostContainer/PostContainer";
 
 export function FeedPage() {
-  const [posts, setPosts] = useState([]);
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const token = localStorage.getItem("token");
-    const loggedIn = token !== null;
-    if (loggedIn) {
-      getPosts(token)
-        .then((data) => {
-          setPosts(data.posts);
-          localStorage.setItem("token", data.token);
-        })
-        .catch((err) => {
-          console.error(err);
-          navigate("/login");
-        });
-    } else {
-      navigate("/login");
-      return;
-    }
-  }, [navigate]);
-
   return (
     <>
       <Header></Header>
       <h2>Posts</h2>
-      <div className="feed" role="feed">
-        {posts.map((post) => (
-          <Post post={post} key={post._id} />
-        ))}
-      </div>
+      <div className="NewPost">{/* <NewPost></NewPost> */}</div>
+      <PostContainer></PostContainer>
     </>
   );
 }
